+++ Source/WebCore/page/scrolling/nicosia/ScrollingTreeFixedNodeNicosia.cpp
@@ -32,7 +32,6 @@
 #if ENABLE(ASYNC_SCROLLING) && USE(NICOSIA)
 
 #include "Logging.h"
-#include "NicosiaPlatformLayer.h"
 #include "ScrollingStateFixedNode.h"
 #include <wtf/text/TextStream.h>
