+++ Source/WebCore/page/scrolling/nicosia/ScrollingTreeFixedNodeNicosia.h
@@ -30,7 +30,7 @@
 
 #if ENABLE(ASYNC_SCROLLING) && USE(NICOSIA)
 
-#include "NicosiaPlatformLayer.h"
+#include "NicosiaCompositionLayer.h"
 #include "ScrollingTreeFixedNode.h"
 #include <wtf/RefPtr.h>
