+++ Source/WebCore/page/scrolling/nicosia/ScrollingTreePositionedNodeNicosia.cpp
@@ -32,7 +32,6 @@
 #if ENABLE(ASYNC_SCROLLING) && USE(NICOSIA)
 
 #include "Logging.h"
-#include "NicosiaPlatformLayer.h"
 #include "ScrollingStatePositionedNode.h"
 
 namespace WebCore {
