+++ Source/WebCore/page/scrolling/nicosia/ScrollingTreePositionedNodeNicosia.h
@@ -30,7 +30,7 @@
 
 #if ENABLE(ASYNC_SCROLLING) && USE(NICOSIA)
 
-#include "NicosiaPlatformLayer.h"
+#include "NicosiaCompositionLayer.h"
 #include "ScrollingTreePositionedNode.h"
 
 namespace WebCore {
