+++ Source/WebCore/platform/TextureMapper.cmake
@@ -90,41 +90,37 @@ if (USE_COORDINATED_GRAPHICS)
         "${WEBCORE_DIR}/platform/graphics/nicosia/texmap"
     )
     list(APPEND WebCore_SOURCES
+        platform/graphics/nicosia/NicosiaBackingStore.cpp
         platform/graphics/nicosia/NicosiaBuffer.cpp
+        platform/graphics/nicosia/NicosiaContentLayer.cpp
+        platform/graphics/nicosia/NicosiaImageBacking.cpp
+        platform/graphics/nicosia/NicosiaImageBackingStore.cpp
         platform/graphics/nicosia/NicosiaImageBufferPipe.cpp
         platform/graphics/nicosia/NicosiaPaintingContext.cpp
         platform/graphics/nicosia/NicosiaPaintingEngine.cpp
         platform/graphics/nicosia/NicosiaPaintingEngineBasic.cpp
         platform/graphics/nicosia/NicosiaPaintingEngineThreaded.cpp
-        platform/graphics/nicosia/NicosiaPlatformLayer.cpp
         platform/graphics/nicosia/NicosiaScene.cpp
         platform/graphics/nicosia/NicosiaSceneIntegration.cpp
 
         platform/graphics/nicosia/cairo/NicosiaCairoOperationRecorder.cpp
         platform/graphics/nicosia/cairo/NicosiaPaintingContextCairo.cpp
-
-        platform/graphics/nicosia/texmap/NicosiaBackingStoreTextureMapperImpl.cpp
-        platform/graphics/nicosia/texmap/NicosiaCompositionLayerTextureMapperImpl.cpp
-        platform/graphics/nicosia/texmap/NicosiaContentLayerTextureMapperImpl.cpp
-        platform/graphics/nicosia/texmap/NicosiaImageBackingStore.cpp
-        platform/graphics/nicosia/texmap/NicosiaImageBackingTextureMapperImpl.cpp
     )
     list(APPEND WebCore_PRIVATE_FRAMEWORK_HEADERS
         page/scrolling/nicosia/ScrollingTreeFixedNodeNicosia.h
         page/scrolling/nicosia/ScrollingTreeStickyNodeNicosia.h
 
         platform/graphics/nicosia/NicosiaAnimatedBackingStoreClient.h
+        platform/graphics/nicosia/NicosiaBackingStore.h
         platform/graphics/nicosia/NicosiaBuffer.h
+        platform/graphics/nicosia/NicosiaCompositionLayer.h
+        platform/graphics/nicosia/NicosiaContentLayer.h
+        platform/graphics/nicosia/NicosiaImageBacking.h
+        platform/graphics/nicosia/NicosiaImageBackingStore.h
         platform/graphics/nicosia/NicosiaPaintingEngine.h
         platform/graphics/nicosia/NicosiaPlatformLayer.h
         platform/graphics/nicosia/NicosiaScene.h
         platform/graphics/nicosia/NicosiaSceneIntegration.h
-
-        platform/graphics/nicosia/texmap/NicosiaBackingStoreTextureMapperImpl.h
-        platform/graphics/nicosia/texmap/NicosiaCompositionLayerTextureMapperImpl.h
-        platform/graphics/nicosia/texmap/NicosiaContentLayerTextureMapperImpl.h
-        platform/graphics/nicosia/texmap/NicosiaImageBackingStore.h
-        platform/graphics/nicosia/texmap/NicosiaImageBackingTextureMapperImpl.h
     )
 else ()
     list(APPEND WebCore_SOURCES
@@ -135,7 +131,7 @@ endif ()
 
 if (ENABLE_WEBGL)
     list(APPEND WebCore_SOURCES
-        platform/graphics/nicosia/texmap/NicosiaGCGLANGLELayer.cpp
+        platform/graphics/nicosia/NicosiaGCGLANGLELayer.cpp
 
         platform/graphics/texmap/GraphicsContextGLTextureMapperANGLENicosia.cpp
     )
